requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.33.4
botocore==1.33.4
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decouple==0.0.7
dj-database-url==2.1.0
Django==4.2.5
django-extra-views==0.14.0
django-multiupload==0.6.1
django-on-heroku==1.1.2
django-storages==1.14.2
exceptiongroup==1.1.3
flower==2.0.1
gunicorn==21.2.0
humanize==4.9.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.3.4
packaging==23.2
pika==1.3.2
Pillow==10.0.1
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pytest==7.4.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
